Set settings that aren't already present
The entrypoint's setting() function only edits settings that are already there in the config files. If the setting is not present, nothing happens (I discovered this while setting the "read_only" setting in neo4j.properties but it is true for other settings too based on my experiments. This changes the setting() function to add the setting to the end of the file if it is not already present in the file.